# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Teleops IOT server API",
        "title": "Teleops  API",
        "contact": {
            "name": "API Support",
            "email": "info@teleops.io"
        },
        "version": "2.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/data/": {
            "get": {
                "description": "Get all Devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data End Points"
                ],
                "summary": "Get all Data",
                "operationId": "get-all-data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start Date",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End Date",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/data/create": {
            "post": {
                "description": "Save device data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data End Points"
                ],
                "summary": "Save device data",
                "operationId": "create-data",
                "parameters": [
                    {
                        "description": "Data Details",
                        "name": "Data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Data"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/devices/": {
            "get": {
                "description": "Get all Devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices End Points"
                ],
                "summary": "Get all Devices",
                "operationId": "get-item-by-int",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/devices/create": {
            "post": {
                "description": "Create a device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices End Points"
                ],
                "summary": "Create a device",
                "operationId": "create-device",
                "parameters": [
                    {
                        "description": "Device Details",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/devices/{id}": {
            "get": {
                "description": "Get device by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices End Points"
                ],
                "summary": "Get device by id",
                "operationId": "get-one-device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete device by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices End Points"
                ],
                "summary": "Delete device by id",
                "operationId": "delete-one-device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update device by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices End Points"
                ],
                "summary": "Update device by id",
                "operationId": "update-one-device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Device Details",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Data": {
            "type": "object",
            "required": [
                "device_address"
            ],
            "properties": {
                "created_by_id": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "datetime": {
                                "type": "string"
                            },
                            "geolocation": {
                                "type": "object",
                                "properties": {
                                    "lat": {
                                        "type": "string"
                                    },
                                    "lon": {
                                        "type": "string"
                                    }
                                }
                            },
                            "processed_value": {
                                "type": "string"
                            },
                            "raw_value": {
                                "type": "string"
                            },
                            "unit_string": {
                                "type": "string"
                            },
                            "units": {
                                "type": "string"
                            }
                        }
                    }
                },
                "datetime": {
                    "type": "string"
                },
                "device_address": {
                    "type": "string"
                },
                "device_alias": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "group_name": {
                    "type": "string"
                },
                "has_geolocation_data": {
                    "type": "boolean"
                },
                "is_deployed": {
                    "type": "boolean"
                },
                "location_id": {
                    "type": "string"
                },
                "location_name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "owner_name": {
                    "type": "string"
                },
                "physical_location": {
                    "type": "string"
                },
                "sensor_class_id": {
                    "type": "string"
                },
                "sensor_class_name": {
                    "type": "string"
                },
                "sensor_type_id": {
                    "type": "string"
                },
                "sensor_type_name": {
                    "type": "string"
                }
            }
        },
        "models.Device": {
            "type": "object",
            "required": [
                "device_address"
            ],
            "properties": {
                "client_details": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "can_get_sms": {
                                "type": "boolean"
                            },
                            "email": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "phone": {
                                "type": "string"
                            }
                        }
                    }
                },
                "created_by_email": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "string"
                },
                "device_address": {
                    "type": "string"
                },
                "device_alias": {
                    "type": "string"
                },
                "geolocation": {
                    "type": "object",
                    "properties": {
                        "lat": {
                            "type": "string"
                        },
                        "lon": {
                            "type": "string"
                        }
                    }
                },
                "group_id": {
                    "type": "string"
                },
                "group_name": {
                    "type": "string"
                },
                "has_geolocation_data": {
                    "type": "boolean"
                },
                "is_deployed": {
                    "type": "boolean"
                },
                "location_id": {
                    "type": "string"
                },
                "location_name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "owner_name": {
                    "type": "string"
                },
                "physical_location": {
                    "type": "string"
                },
                "sensor_class_id": {
                    "type": "string"
                },
                "sensor_class_name": {
                    "type": "string"
                },
                "sensor_type_id": {
                    "type": "string"
                },
                "sensor_type_name": {
                    "type": "string"
                }
            }
        },
        "utils.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}